Organise classes into packages
All classes are currently in a single mickey package, making the
codebase harder to navigate as it grows.

Let's organise classes into:
* mickey.task - task classes
* mickey.util - utility classes (UI, Parser, FileSaver)

This improves maintainability and follows better OOP organisation.
