+++ src/main/java/mickey/task/Event.java
@@ -1,4 +1,4 @@
-package mickey;
+package mickey.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
