+++ src/main/java/mickey/task/TaskList.java
@@ -1,4 +1,4 @@
-package mickey;
+package mickey.task;
 
 import java.util.ArrayList;
